% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_summary.R
\name{combineTaxonomy_ByGroup}
\alias{combineTaxonomy_ByGroup}
\title{Title}
\usage{
combineTaxonomy_ByGroup(taxonomy_table, metadata, group)
}
\arguments{
\item{taxonomy_table}{tibble}

\item{metadata}{tibble}

\item{group}{group}
}
\value{
tibble
}
\description{
Title
}
\examples{
data(metadata)
Class_otu <- microbNIE:::Class_otu
results_comByGroup <- combineTaxonomy_ByGroup(Class_otu, metadata, site)

}
