% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{ggplot_composition}
\alias{ggplot_composition}
\title{Title}
\usage{
ggplot_composition(ggtable_comp)
}
\arguments{
\item{ggtable_comp}{tibble}
}
\value{
ggobject
}
\description{
Title
}
\examples{
ggtable_Comp <- microbNIE:::ggtable_Comp
ggresults <- ggplot_composition(ggtable_Comp)
}
