% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{ggtable_composition}
\alias{ggtable_composition}
\title{Title}
\usage{
ggtable_composition(combined_tax, Num)
}
\arguments{
\item{combined_tax}{tibble}

\item{Num}{integer}
}
\value{
tibble
}
\description{
Title
}
\examples{
combined_Table <- microbNIE:::combined_Table
results_ggtable <- ggtable_composition(combined_Table, 10)
}
