% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{the group information about the sample.}
\format{
A tibble with 126 rows and 2 variables:
\describe{
\item{sample}{sample name}
\item{site}{small group}
...
}
}
\source{
\url{http://www.nowhere1 <- .info/}
}
\usage{
metadata
}
\description{
A dataset containing the sample names and group information
}
\keyword{datasets}
