% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{otu}
\alias{otu}
\title{the results  otu txt table from qiime1}
\format{
A tibble with 3000 rows and 128 variables:
\describe{
\item{otu_id}{the otu id}
\item{23}{the sample name}
...
}
}
\source{
\url{http://www.nowhere2.info/}
}
\usage{
otu
}
\description{
A dataset containing the sample names and otu number
}
\keyword{datasets}
