% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{read_qiime1}
\alias{read_qiime1}
\title{Title read raw otu table from biom-convert in qiime1}
\usage{
read_qiime1(Path, rename = FALSE)
}
\arguments{
\item{Path}{character}

\item{rename}{logic}
}
\value{
tibble
}
\description{
Title read raw otu table from biom-convert in qiime1
}
\examples{
PATH <- system.file("extdata","otu_example.txt", package = "microbNIE", mustWork = TRUE)
example_read <- read_qiime1(PATH, rename = TRUE)
}
