% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_summary.R
\name{split_taxonomy}
\alias{split_taxonomy}
\title{Title}
\usage{
split_taxonomy(otu_table, level)
}
\arguments{
\item{otu_table}{tibble}

\item{level}{level}
}
\value{
tibble
}
\description{
Title
}
\examples{
data(metadata)
data(otu)
results_split_tax <- split_taxonomy(otu, class)
}
